Move PlaneCut and the lone bit of Polygon impl into polygon.rs.
